flaskServer/requirements.txt
psycopg2-binary
flask
streamlit
python-dotenv